Refactoring workspace, workspaceID and blocks completed
