+++ src/common/hooks/index.ts
@@ -1,8 +1,10 @@
 
-export * from './useAddBlock'
+
 export * from './useDeleteBlock'
+export * from './useTagsAndEntities'
 
 export * from './useGetAccountName'
 
 export * from './sandbox'
 export * from './workspaceID'
+export * from './blocks'
