+++ src/common/hooks/sandbox/useGetSandboxData.tsx
@@ -68,6 +68,6 @@ export const useGetSandboxData = (walletData: any) => {
   return { 
     nodeData,
     loading,
-    sandboxData
+    sandboxData,
   }
 }
